Release v0.4.3

+++ contracts/Helpers/TargetContract.sol
@@ -1,18 +1,25 @@
 // SPDX-License-Identifier: Apache-2.0
 pragma solidity ^0.8.0;
 
+/**
+ * @dev sample contract to test interaction + state changes:
+ *      - directly from Universal Profile
+ *      - via KeyManager > UniversalProfile
+ *
+ * also used to test permissions ALLOWEDADDRESS and ALLOWEDSTANDARDS
+ */
 contract TargetContract {
-
     /* solhint-disable */
-    uint number = 5;
+    uint256 number = 5;
     string name = "Simple Contract Name";
+
     /* solhint-enable */
 
-    function getNumber() public view returns (uint) {
+    function getNumber() public view returns (uint256) {
         return number;
     }
 
-    function setNumber(uint _newNumber) public {
+    function setNumber(uint256 _newNumber) public {
         number = _newNumber;
     }
 
@@ -27,5 +34,4 @@ contract TargetContract {
     function revertCall() public pure {
         revert("TargetContract:revertCall: this function has reverted!");
     }
-    
 }
